+++ config.xml
@@ -81,11 +81,6 @@
     <plugin name="cordova-plugin-splashscreen" spec="5.0.2" />
     <plugin name="cordova-plugin-ionic-webview" spec="^2.0.0" />
     <plugin name="cordova-plugin-ionic-keyboard" spec="^2.0.5" />
-    <plugin name="cordova-plugin-googleplus" spec="5.3.2">
-        <variable name="REVERSED_CLIENT_ID" value="com.googleusercontent.apps.96328966461-p4bltjv1c7pd5qvhv44mvi68servchs4" />
-        <variable name="WEB_APPLICATION_CLIENT_ID" value="http://localhost:8000/" />
-    </plugin>
-    <plugin name="cordova-plugin-sqlite-2" spec="1.0.5" />
     <plugin name="cordova-plugin-statusbar" spec="^2.4.2" />
     <plugin name="cordova-android-support-gradle-release" spec="^2.0.1">
         <variable name="ANDROID_SUPPORT_VERSION" value="27.+" />
@@ -96,6 +91,11 @@
         <variable name="FACEBOOK_ANDROID_SDK_VERSION" value="4.38.1" />
     </plugin>
     <plugin name="cordova-sqlite-storage" spec="2.6.0" />
+    <plugin name="cordova-background-geolocation-lt" spec="^2.14.2">
+        <variable name="LICENSE" value="YOUR_LICENSE_KEY" />
+        <variable name="GOOGLE_API_VERSION" value="16.0.0" />
+        <variable name="APPCOMPAT_VERSION" value="27.1.1" />
+    </plugin>
     <engine name="browser" spec="5.0.4" />
-    <engine name="android" spec="7.1.4" />
+    <engine name="android" spec="^7.1.4" />
 </widget>
