+++ package.json
@@ -28,16 +28,16 @@
     "@ionic-native/splash-screen": "^4.18.0",
     "@ionic-native/status-bar": "^4.18.0",
     "@ionic/storage": "^2.2.0",
-    "cordova-android": "7.1.4",
+    "cordova-android": "^7.1.4",
     "cordova-android-support-gradle-release": "^2.0.1",
+    "cordova-background-geolocation-lt": "^2.14.2",
     "cordova-browser": "5.0.4",
+    "cordova-plugin-background-fetch": "^5.4.1",
     "cordova-plugin-device": "^2.0.2",
     "cordova-plugin-facebook4": "4.0.0",
-    "cordova-plugin-googleplus": "5.3.2",
     "cordova-plugin-ionic-keyboard": "^2.1.3",
     "cordova-plugin-ionic-webview": "^2.2.5",
     "cordova-plugin-splashscreen": "^5.0.2",
-    "cordova-plugin-sqlite-2": "1.0.5",
     "cordova-plugin-statusbar": "^2.4.2",
     "cordova-plugin-whitelist": "^1.3.3",
     "cordova-sqlite-storage": "2.6.0",
@@ -50,7 +50,7 @@
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "^3.2.1",
+    "@ionic/app-scripts": "3.2.1",
     "typescript": "^2.7.2"
   },
   "description": "An Ionic project",
@@ -62,11 +62,6 @@
       "cordova-plugin-splashscreen": {},
       "cordova-plugin-ionic-webview": {},
       "cordova-plugin-ionic-keyboard": {},
-      "cordova-plugin-googleplus": {
-        "REVERSED_CLIENT_ID": "com.googleusercontent.apps.96328966461-p4bltjv1c7pd5qvhv44mvi68servchs4",
-        "WEB_APPLICATION_CLIENT_ID": "http://localhost:8000/"
-      },
-      "cordova-plugin-sqlite-2": {},
       "cordova-android-support-gradle-release": {
         "ANDROID_SUPPORT_VERSION": "27.+"
       },
@@ -75,7 +70,12 @@
         "APP_NAME": "progetto_emad_capperfit",
         "FACEBOOK_ANDROID_SDK_VERSION": "4.38.1"
       },
-      "cordova-sqlite-storage": {}
+      "cordova-sqlite-storage": {},
+      "cordova-background-geolocation-lt": {
+        "LICENSE": "YOUR_LICENSE_KEY",
+        "GOOGLE_API_VERSION": "16.0.0",
+        "APPCOMPAT_VERSION": "27.1.1"
+      }
     },
     "platforms": [
       "browser",
