+++ src/index.html
@@ -17,6 +17,7 @@
 
   <!-- cordova.js required for cordova apps (remove if not needed) -->
   <script src="cordova.js"></script>
+  <script src="http://maps.googleapis.com/maps/api/js?key=AIzaSyAtIlYf4cYwSkeXQJXldcFtT5xcOqDXcZw"></script>
 
   <!-- un-comment this code to enable service worker
   <script>
