+++ src/pages/tabs/tabs.scss
@@ -1,0 +1,3 @@
+page-tabs {
+
+}
