+++ src/pages/user-info/user-info.html
@@ -1,0 +1,23 @@
+<!--
+  Generated template for the UserInfoPage page.
+
+  See http://ionicframework.com/docs/components/#navigation for more info on
+  Ionic pages and navigation.
+-->
+<ion-header>
+
+  <ion-navbar>
+    <ion-title>UserInfo</ion-title>
+  </ion-navbar>
+
+</ion-header>
+
+
+<ion-content padding>
+  <p> Contenuto profilo in arrivo </p>
+
+  <button ion-button block (click)="logout();">Test logout</button>
+
+  <p>Test user ID utente:</p>
+  <p id="userID">{{userID}}</p>
+</ion-content>
