+++ src/pages/user-info/user-info.scss
@@ -1,0 +1,3 @@
+page-user-info {
+
+}
