+++ src/pages/user-profile/user-profile.html
@@ -1,0 +1,16 @@
+<ion-header>
+  <ion-navbar>
+    <ion-title style="text-align: center;">
+      HomePage
+    </ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content padding>
+  <p> Contenuto home in arrivo </p>
+
+  <button ion-button block (click)="logout();">Test logout</button>
+
+  <p>User ID utente:</p>
+  <p id="userID">{{userID}}</p>
+</ion-content>
