+++ src/pages/user-profile/user-profile.scss
@@ -1,0 +1,3 @@
+page-user-profile {
+
+}
